backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic-settings==2.6.1
python-multipart==0.0.17
pandas==2.2.3
openpyxl==3.1.5
redis==5.2.1
apscheduler==3.10.4
httpx==0.28.1
python-dateutil==2.9.0.post0
boto3==1.35.76
python-jose[cryptography]==3.3.0
bcrypt>=4.0.1,<5
